Correct terminology in README re: logs_to_follow
The README stated that `node['le']['logs_to_follow']` is

> “An array of logs to follow or a hash of arrays”

but the other docs clearly contradict this, as does the actual code of
the cookbook itself.

This language clarifies the actual functionality.

It should be noted that Chef attributes are remarkably poor at handling
assigning an attribute to be an array, and do much better when treated
as hash-like.